Keep governance benchmark outside scanner package
