+++ src/agents_shipgate/cli/fixture.py
@@ -8,6 +8,7 @@ from __future__ import annotations
 import json
 import shutil
 import subprocess
+from collections.abc import Mapping
 from pathlib import Path
 
 import typer
@@ -237,15 +238,14 @@ def _run_ai_generated_refund_pr_fixture(
     shutil.rmtree(target / "_head", ignore_errors=True)
 
     try:
-        _git(target, "init", "-q", "-b", "main")
-        _git(target, "config", "user.email", "fixture@example.com")
-        _git(target, "config", "user.name", "Agents Shipgate Fixture")
-        _git(target, "add", ".")
-        _git(target, "commit", "-q", "-m", "base support agent")
-        _git(target, "update-ref", "refs/remotes/origin/main", "HEAD")
-        (target / "tools.json").write_text(head_payload, encoding="utf-8")
-        _git(target, "add", "tools.json")
-        _git(target, "commit", "-q", "-m", "codex adds refund tool")
+        materialize_git_pr_fixture(
+            target,
+            head_files={"tools.json": head_payload},
+            user_email="fixture@example.com",
+            user_name="Agents Shipgate Fixture",
+            base_commit_message="base support agent",
+            head_commit_message="codex adds refund tool",
+        )
     except subprocess.CalledProcessError as exc:
         typer.echo(f"Fixture {name!r} git setup failed: {exc}", err=True)
         raise typer.Exit(4) from exc
@@ -307,14 +307,15 @@ def _finish_fixture_copy(
     shutil.rmtree(workdir, ignore_errors=True)
 
 
-def _git(cwd: Path, *args: str) -> None:
-    subprocess.run(
+def _git(cwd: Path, *args: str) -> str:
+    completed = subprocess.run(
         ["git", *args],
         cwd=cwd,
         check=True,
         capture_output=True,
         text=True,
     )
+    return completed.stdout.strip()
 
 
 def _resolve_fixture(name: str) -> Path:
@@ -326,3 +327,63 @@ def _resolve_fixture(name: str) -> Path:
     except FixtureNotFoundError as exc:
         typer.echo(str(exc), err=True)
         raise typer.Exit(2) from exc
+
+
+def materialize_git_pr_fixture(
+    repo: Path,
+    *,
+    base_files: Mapping[str, str | None] | None = None,
+    head_files: Mapping[str, str | None],
+    user_email: str,
+    user_name: str,
+    base_commit_message: str,
+    head_commit_message: str,
+) -> None:
+    """Create a deterministic base/head git fixture with origin/main.
+
+    The fixture CLI and internal benchmark scripts both need the same minimal
+    PR-shaped history. Keeping the git plumbing here avoids parallel copies of
+    the audited subprocess call site.
+    """
+
+    if base_files:
+        _apply_fixture_files(repo, base_files)
+    _git(repo, "init", "-q", "-b", "main")
+    _git(repo, "config", "user.email", user_email)
+    _git(repo, "config", "user.name", user_name)
+    _git(repo, "add", ".")
+    _git(repo, "commit", "-q", "-m", base_commit_message)
+    _git(repo, "update-ref", "refs/remotes/origin/main", "HEAD")
+    _apply_fixture_files(repo, head_files)
+    _git(repo, "add", ".")
+    _git(repo, "commit", "-q", "-m", head_commit_message)
+
+
+def run_fixture_git(repo: Path, *args: str) -> str:
+    """Run the fixture git helper and return stripped stdout."""
+
+    return _git(repo, *args)
+
+
+def _apply_fixture_files(repo: Path, files: Mapping[str, str | None]) -> None:
+    for relative, content in sorted(files.items()):
+        path = repo / relative
+        if content is None:
+            if path.exists():
+                path.unlink()
+            _prune_empty_parents(path.parent, repo)
+            continue
+        path.parent.mkdir(parents=True, exist_ok=True)
+        path.write_text(content, encoding="utf-8")
+
+
+def _prune_empty_parents(path: Path, root: Path) -> None:
+    while path != root and path.exists():
+        try:
+            path.rmdir()
+        except OSError:
+            return
+        path = path.parent
+
+
+__all__ = ["fixture_app", "materialize_git_pr_fixture", "run_fixture_git"]
